PythonApp/requirements.txt
numpy==1.23.4
pandas==1.5.1
Pillow==9.5.0
psycopg2==2.9.6
PyQt5==5.15.9
python-dotenv==1.0.0
requests==2.28.1
